refactor(pwa): centralize header/footer in route-group layouts (persistent chrome)
Each page rendered its own header/footer ad hoc, so the footer was missing on
"Mes voyages", the trip view and "Nouveau voyage", and absent entirely from the
loading / "Voyage introuvable" states and the 404/500 error pages (recette #649).
Centralise the chrome so the same header + the homepage footer (which carries the
OSM attribution) appear everywhere; only the content varies per page.

- New <SiteChrome> (header + LandingFooter, content = children); the `app`
  variant also mounts the HelpModal so the help button works app-wide.
- Route groups: (app)/layout (authenticated TopBar) wraps Mes voyages, Nouveau
  voyage, the trip view + its loading / "Voyage introuvable" states, and account
  settings; (public)/layout (PublicTopBar) wraps FAQ, legal, privacy. URLs are
  unchanged (the parentheses are stripped from the path).
- TripPlanner no longer renders its own TopBar / HelpModal / mini-footer; the
  home dashboard and the error pages reuse <SiteChrome variant="app">.
- Exceptions kept by design: the landing (transparent hero header), the login /
  verify auth pages (focused minimal chrome) and the shared /s/[code] page
  (read-only SharedTopBar).

Refs #649
