+++ pwa/src/app/(app)/trips/[id]/page.tsx
