+++ pwa/src/app/(app)/trips/[id]/trip-page.tsx
