+++ pwa/src/app/(public)/faq/page.tsx
@@ -1,0 +1,100 @@
+import Link from "next/link";
+import { ArrowLeft } from "lucide-react";
+import { getTranslations } from "next-intl/server";
+import { FaqAccordion, type FaqCategory } from "@/components/faq-accordion";
+
+export default async function FaqPage() {
+  const t = await getTranslations("faq");
+
+  const categories: FaqCategory[] = [
+    {
+      id: "project",
+      label: t("categoryProject"),
+      items: [
+        { question: t("q1"), answer: t("a1") },
+        { question: t("q2"), answer: t("a2") },
+        { question: t("q3"), answer: t("a3") },
+        { question: t("q4"), answer: t("a4") },
+      ],
+    },
+    {
+      id: "how-it-works",
+      label: t("categoryHowItWorks"),
+      items: [
+        { question: t("q5"), answer: t("a5") },
+        { question: t("q6"), answer: t("a6") },
+        { question: t("q7"), answer: t("a7") },
+      ],
+    },
+    {
+      id: "access",
+      label: t("categoryAccess"),
+      items: [
+        { question: t("q8"), answer: t("a8") },
+        { question: t("q9"), answer: t("a9") },
+      ],
+    },
+    {
+      id: "ai",
+      label: t("categoryAi"),
+      items: [
+        { question: t("q10"), answer: t("a10") },
+        { question: t("q11"), answer: t("a11") },
+      ],
+    },
+  ];
+
+  return (
+    <main
+      className="flex-1 w-full max-w-5xl mx-auto px-4 md:px-6 py-12"
+      data-testid="faq-page"
+    >
+      {/* Back link */}
+      <Link
+        href="/"
+        className="inline-flex items-center gap-1.5 text-sm text-muted-foreground hover:text-foreground mb-8 transition-colors"
+        data-testid="faq-back-link"
+      >
+        <ArrowLeft className="h-4 w-4" aria-hidden="true" />
+        {t("backToHome")}
+      </Link>
+
+      {/* Header */}
+      <div className="mb-10">
+        <h1 className="text-3xl font-bold tracking-tight mb-2">{t("title")}</h1>
+        <p className="text-muted-foreground">{t("description")}</p>
+      </div>
+
+      <div className="grid gap-10 md:grid-cols-[16rem_1fr]">
+        {/* Sections rail (sticky on desktop) */}
+        <nav
+          aria-label={t("sectionsLabel")}
+          className="md:sticky md:top-12 md:self-start"
+          data-testid="faq-sections-rail"
+        >
+          <p className="text-xs font-semibold uppercase tracking-wide text-muted-foreground mb-3">
+            {t("sectionsLabel")}
+          </p>
+          <ul className="space-y-2 text-sm border-l border-border">
+            {categories.map((category) => (
+              <li key={category.id}>
+                <a
+                  href={`#faq-category-${category.id}`}
+                  className="block -ml-px border-l border-transparent pl-4 text-muted-foreground hover:text-foreground hover:border-brand transition-colors"
+                  data-testid={`faq-sections-link-${category.id}`}
+                >
+                  {category.label}
+                </a>
+              </li>
+            ))}
+          </ul>
+        </nav>
+
+        {/* Content */}
+        <div className="min-w-0">
+          <FaqAccordion categories={categories} />
+        </div>
+      </div>
+    </main>
+  );
+}
