+++ pwa/src/app/(public)/legal/page.tsx
@@ -5,8 +5,6 @@ import {
   asParagraphs,
   withContactEmail,
 } from "@/components/legal-page";
-import { PublicTopBar } from "@/components/public-top-bar";
-import { LandingFooter } from "@/components/landing/footer";
 
 const SECTION_IDS = [
   "publisher",
@@ -28,18 +26,14 @@ export default async function LegalNoticePage() {
   });
 
   return (
-    <>
-      <PublicTopBar />
-      <LegalPageLayout
-        title={t("title")}
-        intro={t("intro")}
-        lastUpdated={t("lastUpdated")}
-        tocLabel={t("tocLabel")}
-        backToHome={t("backToHome")}
-        sections={sections}
-        testIdPrefix="legal"
-      />
-      <LandingFooter />
-    </>
+    <LegalPageLayout
+      title={t("title")}
+      intro={t("intro")}
+      lastUpdated={t("lastUpdated")}
+      tocLabel={t("tocLabel")}
+      backToHome={t("backToHome")}
+      sections={sections}
+      testIdPrefix="legal"
+    />
   );
 }
